+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.0 (DEV)
 
 - Add: `*.sp` files support.
+- Add: Convert entire file to a different format.
 - Enh: Minor UI enhancements.
 - Internal: Refactored Drivers, so now it's only one "Supaplex" Driver with
   different file formats.